Adding viewLog and localStorage helpers
Two new files to help the platform manage the Epic view log on the client. One is a set of helper functions for the view log, the other is a set of helper functions to interact with localStorage. Only the former gets exported from this library. Both files ported and refactored from Frontend.