Local dev changes
Recommend yarn link for local dev. This avoids the need to re-add the
module in the consuming project to pick up changes.

Add yarn dev command. This watches for changes and recompiles
automatically. This saves having to manually re-run yarn build when
developing against a local version of the client.